requirements.txt
numpy==1.18.1
matplotlib==3.2.0
requests==2.11.1
tensorflow==2.1.0
h5py==2.10.0
Keras==2.3.1
opencv_contrib_python==4.2.0.32
pandas==1.0.1
Pillow==8.1.1
progressbar33==2.4
wandb==0.10.21
